issue #1 [Fibonacci]: Fix wrong result for 50 and plus index

+++ src/main/java/com/liva/dojo/fibonacci/service/FibonacciService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class FibonacciService {
 
-    public int fibonacci(int i) {
+    public long fibonacci(int i) {
         if(i < 0) {
             throw new IllegalArgumentException("Input must be a non-negative integer");
         }
@@ -15,7 +15,7 @@ public class FibonacciService {
         if (i == 1) {
             return 1;
         }
-        return (int) fibonacci(i - 1) + (int) fibonacci(i - 2);
+        return (long) fibonacci(i - 1) + (long) fibonacci(i - 2);
     }
 
 }
